Weather is now a cog!
And I've checked it works!
