--- DaveBOT/owmaw/__init__.py
@@ -1,1 +1,0 @@
-from .weather import weather
